Converted API from String to Data.Lazy.ByteString.
This is a breaking change! The version number has been bumped to a new
major version number to indicate this! Update your programmes if you're
using the Rabobank library. If you're only using the CLI programme that
comes with this package you don't need to worry.

Reason for the change:

String is the default string format of Haskell, but it is not performing
well, as they are simply large lists ([Char]), especially when large
strings are involved, as they are in this programme. Entire files with
maybe thousands of transactions need to be converted, and this is an
instance where choosing another string type is better. There are 2
alternative, better performing string types in Haskell, namely
ByteString and Text. Text is used mostly when dealing with Unicode, but
since we're dealing with ASCII for now, I'm using ByteString (although
we probably want to support Unicode for just the transaction
descriptions at some point.)